unify virus and hla processing: implement single function for processing inputs to get event posteriors and use it for both hla and virus
