+++ src/calling/haplotypes/virus.rs
@@ -1,5 +1,5 @@
 use crate::calling::haplotypes::haplotypes;
-// use crate::calling::haplotypes::haplotypes::find_similar_lp_haplotypes;
+use crate::calling::haplotypes::haplotypes::{extend_resulting_table, get_event_posteriors};
 // use crate::calling::haplotypes::haplotypes::HaplotypeGraphVirus;
 // use crate::calling::haplotypes::haplotypes::SimilarL;
 use crate::calling::haplotypes::haplotypes::{
@@ -54,102 +54,105 @@ impl Caller {
             self.output_empty_files()?;
             Ok(())
         } else {
-            //FIRST, perform linear program using only nonzero DP variants
-            //prepare haplotype variants
-            let haplotype_variants_all = HaplotypeVariants::new(&mut self.haplotype_variants)?;
 
-            //filter variant calls and haplotype variants
-            let filtered_calls = variant_calls.without_zero_dp();
-            let nonzero_dp_variants: Vec<VariantID> = filtered_calls.keys().cloned().collect();
-            let var_filt_haplotype_variants =
-                haplotype_variants_all.filter_for_variants(&nonzero_dp_variants)?;
+            // //FIRST, perform linear program using only nonzero DP variants
+            // //prepare haplotype variants
+            // let haplotype_variants_all = HaplotypeVariants::new(&mut self.haplotype_variants)?;
 
-            //find identical haplotypes using variants and prepare LP inputs
-            let haplotypes: Vec<Haplotype> = var_filt_haplotype_variants
-                .iter()
-                .next()
-                .unwrap()
-                .1
-                .keys()
-                .cloned()
-                .collect();
-            let candidate_matrix = CandidateMatrix::new(&var_filt_haplotype_variants).unwrap();
-            //generate one map with representative haplotypes as key (required for lp) and one map with all haplotypes as key (required for extension of resulting table)
-            let (identical_haplotypes_map_rep, identical_haplotypes_map) =
-                candidate_matrix.find_identical_haplotypes(haplotypes);
-            let representatives = identical_haplotypes_map_rep.keys().cloned().collect();
-            let repr_haplotype_variants =
-                var_filt_haplotype_variants.filter_for_haplotypes(&representatives)?;
-            let repr_candidate_matrix = CandidateMatrix::new(&repr_haplotype_variants).unwrap();
+            // //filter variant calls and haplotype variants
+            // let filtered_calls = variant_calls.without_zero_dp();
+            // let nonzero_dp_variants: Vec<VariantID> = filtered_calls.keys().cloned().collect();
+            // let var_filt_haplotype_variants =
+            //     haplotype_variants_all.filter_for_variants(&nonzero_dp_variants)?;
 
-            //employ the linear program
-            //note: extension is disabled at the moment. see notes on linear_program function.
-            let lp_haplotypes = haplotypes::linear_program(
-                &self.outcsv,
-                &repr_candidate_matrix,
-                &representatives,
-                &filtered_calls,
-                self.lp_cutoff,
-                self.extend_haplotypes,
-                self.num_extend_haplotypes, //extension functionality for virus case is not recommended for now as it will lead to performance problems.
-                self.num_constraint_haplotypes,
-            )?;
+            // //find identical haplotypes using variants and prepare LP inputs
+            // let haplotypes: Vec<Haplotype> = var_filt_haplotype_variants
+            //     .iter()
+            //     .next()
+            //     .unwrap()
+            //     .1
+            //     .keys()
+            //     .cloned()
+            //     .collect();
+            // let candidate_matrix = CandidateMatrix::new(&var_filt_haplotype_variants).unwrap();
+            // //generate one map with representative haplotypes as key (required for lp) and one map with all haplotypes as key (required for extension of resulting table)
+            // let (identical_haplotypes_map_rep, identical_haplotypes_map) =
+            //     candidate_matrix.find_identical_haplotypes(haplotypes);
+            // let representatives = identical_haplotypes_map_rep.keys().cloned().collect();
+            // let repr_haplotype_variants =
+            //     var_filt_haplotype_variants.filter_for_haplotypes(&representatives)?;
+            // let repr_candidate_matrix = CandidateMatrix::new(&repr_haplotype_variants).unwrap();
 
-            //SECOND, model evaluation using ALL variants but only the LP- selected haplotypes
-            //prepare inputs of model evaluation
-            let hap_filt_haplotype_variants =
-                haplotype_variants_all.filter_for_haplotypes(&lp_haplotypes)?;
-            let model_candidate_matrix = CandidateMatrix::new(&hap_filt_haplotype_variants)?;
+            // //employ the linear program
+            // //note: extension is disabled at the moment. see notes on linear_program function.
+            // let lp_haplotypes = haplotypes::linear_program(
+            //     &self.outcsv,
+            //     &repr_candidate_matrix,
+            //     &representatives,
+            //     &filtered_calls,
+            //     self.lp_cutoff,
+            //     self.extend_haplotypes,
+            //     self.num_extend_haplotypes, //extension functionality for virus case is not recommended for now as it will lead to performance problems.
+            //     self.num_constraint_haplotypes,
+            // )?;
 
-            //compute model
-            let prior = PriorTypes::from_str(&self.prior).unwrap();
-            let upper_bond = NotNan::new(1.0).unwrap();
-            let model = Model::new(
-                Likelihood::new(),
-                Prior::new(prior.clone()),
-                Posterior::new(),
-            );
+            // //SECOND, model evaluation using ALL variants but only the LP- selected haplotypes
+            // //prepare inputs of model evaluation
+            // let hap_filt_haplotype_variants =
+            //     haplotype_variants_all.filter_for_haplotypes(&lp_haplotypes)?;
+            // let model_candidate_matrix = CandidateMatrix::new(&hap_filt_haplotype_variants)?;
 
-            let data = Data::new(model_candidate_matrix.clone(), variant_calls.clone());
+            // //compute model
+            // let prior = PriorTypes::from_str(&self.prior).unwrap();
+            // let upper_bond = NotNan::new(1.0).unwrap();
+            // let model = Model::new(
+            //     Likelihood::new(),
+            //     Prior::new(prior.clone()),
+            //     Posterior::new(),
+            // );
 
-            //marginal computation
-            let computed_model = model.compute_from_marginal(
-                &Marginal::new(
-                    lp_haplotypes.len(),
-                    lp_haplotypes.clone(),
-                    upper_bond,
-                    prior,
-                    None,
-                    false, //equivalence class constraint is currently disabled.
-                    "virus".to_string(),
-                ),
-                &data,
-            );
+            // let data = Data::new(model_candidate_matrix.clone(), variant_calls.clone());
 
-            //find event posteriors
-            let event_posteriors = computed_model.event_posteriors();
+            // //marginal computation
+            // let computed_model = model.compute_from_marginal(
+            //     &Marginal::new(
+            //         lp_haplotypes.len(),
+            //         lp_haplotypes.clone(),
+            //         upper_bond,
+            //         prior,
+            //         None,
+            //         false, //equivalence class constraint is currently disabled.
+            //         "virus".to_string(),
+            //     ),
+            //     &data,
+            // );
 
-            //remove zero densities from the table
-            let mut event_posteriors = Vec::new();
-            computed_model
-                .event_posteriors()
-                .for_each(|(fractions, logprob)| {
-                    if logprob.exp() != 0.0 {
-                        event_posteriors.push((fractions.clone(), logprob.clone()));
-                    }
-                });
+            // //find event posteriors
+            // let event_posteriors = computed_model.event_posteriors();
 
-            // Third, extend the table with identical haplotypes
-            let (new_event_posteriors, all_haplotypes) = extend_resulting_table(
-                &lp_haplotypes,
-                &event_posteriors,
-                &identical_haplotypes_map,
-            )?;
+            // //remove zero densities from the table
+            // let mut event_posteriors = Vec::new();
+            // computed_model
+            //     .event_posteriors()
+            //     .for_each(|(fractions, logprob)| {
+            //         if logprob.exp() != 0.0 {
+            //             event_posteriors.push((fractions.clone(), logprob.clone()));
+            //         }
+            //     });
+
+            // // Third, extend the table with identical haplotypes
+            // let (new_event_posteriors, all_haplotypes) = extend_resulting_table(
+            //     &lp_haplotypes,
+            //     &event_posteriors,
+            //     &identical_haplotypes_map,
+            // )?;
+            let haplotype_variants = HaplotypeVariants::new(&mut self.haplotype_variants)?;
+            let (event_posteriors, all_haplotypes, data) = get_event_posteriors(&haplotype_variants, variant_calls, &"virus", &self.prior, &self.outcsv, self.extend_haplotypes, self.num_extend_haplotypes, self.num_constraint_haplotypes, self.lp_cutoff, false, None)?;
 
             //plot the best solution as final solution plot
-            let (best_fractions, _) = new_event_posteriors.iter().next().unwrap();
-            let candidate_matrix_all = CandidateMatrix::new(
-                &haplotype_variants_all
+            let (best_fractions, _) = event_posteriors.iter().next().unwrap();
+            let candidate_matrix = CandidateMatrix::new(
+                &haplotype_variants
                     .filter_for_haplotypes(&all_haplotypes)
                     .unwrap(),
             )
@@ -166,7 +169,7 @@ impl Caller {
             haplotypes::plot_prediction(
                 &self.outcsv,
                 &"final",
-                &candidate_matrix_all,
+                &candidate_matrix,
                 &all_haplotypes,
                 &data.variant_calls,
                 &best_fractions,
@@ -176,7 +179,7 @@ impl Caller {
             haplotypes::write_results(
                 &self.outcsv,
                 &data,
-                &new_event_posteriors,
+                &event_posteriors,
                 &all_haplotypes,
                 self.prior.clone(),
                 false,
@@ -185,7 +188,7 @@ impl Caller {
             //plot first 10 posteriors of orthanq output
             haplotypes::plot_densities(
                 &self.outcsv,
-                &new_event_posteriors,
+                &event_posteriors,
                 &all_haplotypes,
                 "viral",
             )?;
@@ -221,84 +224,6 @@ impl Caller {
         Ok(())
     }
 }
-fn generate_combinations(
-    expanded_fractions: &Vec<AlleleFreq>,
-    haplotype_indices: &BTreeMap<Haplotype, usize>,
-    identical_haplotypes_map: &BTreeMap<Haplotype, Vec<Haplotype>>,
-) -> Vec<Vec<AlleleFreq>> {
-    let mut result = vec![expanded_fractions.clone()];
-
-    for (cur_haplotype, &idx_current) in haplotype_indices.iter() {
-        let fraction = expanded_fractions[idx_current];
-
-        if fraction > NotNan::new(0.0).unwrap() {
-            if let Some(identical_haplotypes) = identical_haplotypes_map.get(cur_haplotype) {
-                let mut new_combinations = Vec::new();
-
-                for existing_row in &result {
-                    for ident_h in identical_haplotypes {
-                        if ident_h != cur_haplotype {
-                            let idx_ident_h = haplotype_indices.get(ident_h).unwrap();
-                            let mut alt_row = existing_row.clone();
-                            alt_row[*idx_ident_h] = fraction;
-                            alt_row[idx_current] = NotNan::new(0.0).unwrap();
-                            new_combinations.push(alt_row);
-                        }
-                    }
-                }
-
-                result.extend(new_combinations);
-            }
-        }
-    }
-
-    result
-}
-
-fn extend_resulting_table(
-    representatives: &Vec<Haplotype>,
-    event_posteriors: &Vec<(HaplotypeFractions, LogProb)>,
-    identical_haplotypes_map: &BTreeMap<Haplotype, Vec<Haplotype>>,
-) -> Result<(Vec<(HaplotypeFractions, LogProb)>, Vec<Haplotype>)> {
-    let mut all_haplotypes = BTreeSet::new();
-    for haplotype in representatives {
-        all_haplotypes.insert(haplotype.clone());
-        let identical_haplotypes = identical_haplotypes_map.get(haplotype).unwrap();
-        all_haplotypes.extend(identical_haplotypes.clone());
-    }
-    let all_haplotypes: Vec<Haplotype> = all_haplotypes.into_iter().collect();
-
-    let mut new_event_posteriors = Vec::new();
-
-    let haplotype_indices: BTreeMap<Haplotype, usize> = all_haplotypes
-        .iter()
-        .enumerate()
-        .map(|(i, hap)| (hap.clone(), i))
-        .collect();
-
-    for (fractions, logprob) in event_posteriors {
-        let mut expanded_fractions = vec![NotNan::new(0.0).unwrap(); all_haplotypes.len()];
-
-        for (i, &fraction) in fractions.iter().enumerate() {
-            if let Some(&idx) = haplotype_indices.get(&representatives[i]) {
-                expanded_fractions[idx] = fraction;
-            }
-        }
-
-        let all_combinations = generate_combinations(
-            &expanded_fractions,
-            &haplotype_indices,
-            identical_haplotypes_map,
-        );
-
-        for combination in all_combinations {
-            new_event_posteriors.push((HaplotypeFractions(combination), logprob.clone()));
-        }
-    }
-
-    Ok((new_event_posteriors, all_haplotypes))
-}
-
 fn filter_representatives(
     haplotypes: Vec<Haplotype>,
     distance_matrix: DistanceMatrix,
